requirements.txt
gradio>=4.0.0
huggingface_hub>=0.23.0
streamlit>=1.25.0
pillow>=9.0.0
